Split `html.scala` into multiple files
